+++ searchmytraining/src/main/java/com/searchmytraining/controller/FreeLancerController.java
@@ -102,7 +102,7 @@ public class FreeLancerController {
 		{
 			List<StateEntity> states = stateservice.getStates(location.getCity().getState().getCountry().getCountryId());
 			List<CityEntity> cities = cityservice.getCities(location.getCity().getState().getStateId());
-			Long country_value = location.getCity().getState().getCountry().getCountryId();
+			Integer country_value = location.getCity().getState().getCountry().getCountryId();
 			model.addAttribute("country_value",country_value);
 			model.addAttribute("location", location);
 			model.addAttribute("states",new JSONArray(states));
