+++ searchmytraining/src/main/java/com/searchmytraining/dao/StateDAO.java
@@ -7,7 +7,7 @@ import com.searchmytraining.entity.StateEntity;
 public interface StateDAO {
 	
 	public StateEntity getState(Integer id);
-	public List<StateEntity> getStates(Long countryid);
+	public List<StateEntity> getStates(Integer countryid);
 	public List<StateEntity> getAllStates();
 
 }
